component refactoring and tidy up
