+++ src/components/Profile/Edit/TouchableRow/TouchableRow.styles.tsx
@@ -1,0 +1,28 @@
+import { typography } from '@/src/theme/typography';
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  inputContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+  },
+  inputRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    flex: 1,
+  },
+  inputTextContainer: {
+    flexDirection: 'column',
+    marginLeft: 10,
+    flex: 1,
+  },
+  inputLabel: {
+    fontSize: typography.fontSizes.body,
+    fontFamily: typography.fontFamilies.poppins.medium,
+  },
+  inputValue: {
+    fontSize: typography.fontSizes.button,
+    fontFamily: typography.fontFamilies.poppins.medium,
+  },
+});
