+++ src/components/Profile/Edit/TouchableRow/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TouchableRow';
