+++ src/components/Profile/ProfileDescription/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProfileDescription';
